+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Added
 - Tags in ActionSchema
 - Entity name in ActionSchema
+- `Api.done()` function returning null for lack of async support.
 
 ## [1.2.8] - 2017-11-07
 ## Changed