+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Tags in ActionSchema
 - Entity name in ActionSchema
 - `Api.done()` function returning null for lack of async support.
+## Changed
+- Removed empty elements from transport writes
 
 ## [1.2.8] - 2017-11-07
 ## Changed